+++ components/svg/SvgLogo.vue
@@ -1,0 +1,183 @@
+<script setup>
+
+</script>
+
+<template lang="pug">
+svg#holologo(version="1.1", xmlns="http://www.w3.org/2000/svg", xlink="http://www.w3.org/1999/xlink", x="0px", y="0px", viewBox="0 0 198.4 198.4", style="enable-background:new 0 0 198.4 198.4;", space="preserve")
+  g#keys
+    g#black
+      path.st0(d="M151.7,45.2c5.1-9,2.1-20.4-6.8-25.5c-8.9-5.2-20.4-2.1-25.5,6.8c-0.5,0.9-1,1.8-1.3,2.8l-0.1,0.2c-0.4,1-0.7,2.1-0.8,3.2l-11.6,43.7c-1.1,3.2,0.7,6.6,3.9,7.7c2.3,0.8,4.8,0.1,6.4-1.7l33.5-33.9C150.2,47.4,151,46.3,151.7,45.2z")
+      path.st0(d="M153.3,151.7c9,5.1,20.4,2.1,25.5-6.8c5.2-8.9,2.1-20.4-6.8-25.5c-0.9-0.5-1.9-1-2.8-1.3l-0.2,0c-1-0.4-2.1-0.7-3.2-0.8l-43.7-11.6c-3.2-1-6.6,0.7-7.7,3.9c-0.7,2.3-0.1,4.8,1.7,6.4l33.9,33.5C151.1,150.3,152.1,151,153.3,151.7z")
+      path.st0(d="M80.8,172.3c-0.1,10.3,8.2,18.8,18.5,18.9s18.8-8.2,18.9-18.5c0-1.2-0.1-2.3-0.3-3.5V169c-0.2-1.1-0.5-2.1-0.9-3.2l-11.8-43.7c-0.7-3.3-3.9-5.4-7.2-4.7c-2.3,0.5-4.2,2.3-4.7,4.7l-12,46.1C81,169.6,80.8,170.9,80.8,172.3z")
+      path.st0(d="M26.1,80.8C15.8,80.6,7.4,89,7.3,99.3s8.2,18.8,18.5,18.9c1.2,0,2.3-0.1,3.5-0.3h0.2c1.1-0.2,2.1-0.5,3.2-0.9
+      l43.7-11.8c3.3-0.7,5.4-3.9,4.7-7.2c-0.5-2.3-2.3-4.2-4.7-4.7l-46.1-12C28.8,81,27.5,80.8,26.1,80.8z")
+      path.st0(d="M78.6,26.7c-5.2-8.9-16.6-12-25.5-6.8s-12,16.6-6.8,25.5c0.5,0.9,1.1,1.8,1.8,2.6l0.1,0.1
+      c0.7,0.8,1.5,1.6,2.3,2.3l32.1,31.9c2.2,2.5,6.1,2.7,8.6,0.5c1.8-1.6,2.5-4.2,1.7-6.5l-12.6-46C79.9,29.1,79.3,27.9,78.6,26.7z")
+    g#white
+      path.st1(d="M171.7,78.6c8.9-5.2,12-16.6,6.8-25.5s-16.6-12-25.5-6.8c-0.9,0.5-1.7,1.1-2.5,1.8l-0.1,0.1
+      c-0.8,0.7-1.6,1.5-2.3,2.3l-31.9,32.1c-2.5,2.2-2.7,6.1-0.5,8.6c1.6,1.8,4.1,2.5,6.4,1.7l46-12.6
+      C169.3,79.8,170.5,79.3,171.7,78.6z")
+      path.st1(d="M172.3,117.6c10.3,0.1,18.7-8.3,18.7-18.6s-8.3-18.7-18.6-18.7c-1.1,0-2.2,0.1-3.2,0.3H169
+      c-1.1,0.2-2.1,0.5-3.2,0.9l-43.7,11.9c-3.3,0.7-5.4,3.9-4.7,7.2c0.5,2.3,2.3,4.2,4.7,4.7l46.1,12
+      C169.6,117.5,170.9,117.6,172.3,117.6z")
+      path.st1(d="M119.8,171.7c5.2,8.9,16.5,12.1,25.5,7c9-5.1,12.1-16.5,7-25.5c-0.5-1-1.2-1.9-1.9-2.7l-0.1-0.1
+      c-0.7-0.8-1.5-1.6-2.3-2.3l-32.1-31.9c-2.2-2.5-6.1-2.7-8.6-0.4c-1.8,1.6-2.4,4.1-1.7,6.4l12.6,46
+      C118.6,169.3,119.1,170.5,119.8,171.7z")
+      path.st1(d="M46.7,153.3c-5.1,9-2.1,20.4,6.8,25.5C62.5,184,74,181,79.1,172c0.5-0.9,1-1.9,1.3-2.8V169
+      c0.4-1,0.7-2.1,0.8-3.2l11.6-43.7c1-3.2-0.7-6.6-3.9-7.7c-2.3-0.7-4.8-0.1-6.4,1.7L49,150.1C48.1,151.1,47.4,152.1,46.7,153.3z")
+      path.st1(d="M26.7,119.8c-8.9,5.2-11.9,16.6-6.7,25.5c5.2,8.9,16.6,11.9,25.5,6.7c0.9-0.5,1.7-1.1,2.4-1.7l0.1-0.1
+      c0.8-0.7,1.6-1.5,2.3-2.3l31.9-32.1c2.5-2.3,2.7-6.1,0.4-8.6c-1.6-1.8-4.1-2.5-6.4-1.7l-46,12.6C29.1,118.6,27.9,119.1,26.7,119.8
+      z")
+      path.st1(d="M45.2,46.7c-8.9-5.1-20.4-2.1-25.5,6.8S17.5,74,26.5,79.1c0.9,0.5,1.8,1,2.8,1.3h0.2c1,0.4,2.1,0.7,3.2,0.8
+      l43.7,11.7c3.2,1.1,6.6-0.7,7.7-3.9c0.8-2.3,0.1-4.8-1.7-6.4l-34-33.5C47.4,48.2,46.3,47.4,45.2,46.7z")
+      path.st1(d="M117.6,26.1c0.1-10.3-8.3-18.7-18.6-18.7S80.3,15.7,80.3,26c0,1.1,0.1,2.2,0.3,3.2v0.2
+      c0.2,1.1,0.5,2.1,0.9,3.2l11.9,43.7c0.7,3.3,3.9,5.4,7.2,4.7c2.3-0.5,4.2-2.3,4.7-4.7l12-46.1C117.5,28.8,117.6,27.5,117.6,26.1z")
+  g#notes
+    g#Gs
+      path.st2(d="M51.8,46.8c-14.3-14.3,7.3-36,21.7-21.6S66.2,61.1,51.8,46.8z")
+      g
+        path.st3(d="M57.7,40.7c-0.9,0-1.6-0.1-2.3-0.4s-1.2-0.6-1.7-1.1c-0.5-0.5-0.8-1-1-1.6c-0.2-0.6-0.4-1.3-0.4-2.1
+        c0-1,0.2-1.9,0.6-2.7c0.4-0.8,1-1.4,1.8-1.8s1.7-0.7,2.8-0.7c0.4,0,0.7,0,1,0.1c0.3,0,0.6,0.1,0.9,0.2c0.3,0.1,0.6,0.2,0.8,0.2
+        c0.3,0.1,0.5,0.2,0.7,0.3l-0.2,2.1c-0.4-0.2-0.8-0.4-1.1-0.5c-0.4-0.1-0.7-0.2-1.1-0.3s-0.6-0.1-0.9-0.1c-0.4,0-0.8,0.1-1.2,0.2
+        c-0.4,0.1-0.7,0.3-1,0.6s-0.5,0.6-0.6,1c-0.2,0.4-0.2,0.9-0.2,1.4c0,0.5,0.1,0.9,0.2,1.3s0.3,0.8,0.6,1.1s0.6,0.5,1,0.7
+        c0.4,0.2,0.9,0.3,1.4,0.3c0.3,0,0.6,0,0.8,0s0.5-0.1,0.8-0.1s0.5-0.1,0.7-0.2c0.2-0.1,0.5-0.2,0.7-0.3l0.3,2.1
+        c-0.2,0.1-0.4,0.1-0.6,0.2c-0.2,0.1-0.5,0.1-0.8,0.2s-0.6,0.1-1,0.1C58.4,40.7,58.1,40.7,57.7,40.7z M59,40.2v-4.9h2.2v4.8
+        L59,40.2z")
+        path.st3(d="M62.2,38.7v-1.5h8.7v1.5H62.2z M62.9,34.6v-1.5h8.7v1.5H62.9z M63.5,40.5c0.1-0.5,1.6-9,1.7-9.3h1.6
+        c-0.1,0.3-1.6,8.8-1.7,9.3H63.5z M67,40.5c0.1-0.5,1.6-9,1.7-9.3h1.6c-0.1,0.3-1.6,8.8-1.7,9.3H67z")
+    g#G
+      path.st4(d="M32,77.5c-19.6-5.2-11.7-34.8,7.9-29.6S51.6,82.7,32,77.5z")
+      path.st3(d="M36.1,68c-0.9,0-1.6-0.1-2.3-0.4s-1.2-0.6-1.7-1.1c-0.5-0.5-0.8-1-1-1.6c-0.2-0.6-0.4-1.3-0.4-2.1
+      c0-1,0.2-1.9,0.6-2.7c0.4-0.8,1-1.4,1.8-1.8s1.7-0.7,2.8-0.7c0.4,0,0.7,0,1,0.1c0.3,0,0.6,0.1,0.9,0.2c0.3,0.1,0.6,0.2,0.8,0.2
+      c0.3,0.1,0.5,0.2,0.7,0.3l-0.2,2.1c-0.4-0.2-0.8-0.4-1.1-0.5c-0.4-0.1-0.7-0.2-1.1-0.3s-0.6-0.1-0.9-0.1c-0.4,0-0.8,0.1-1.2,0.2
+      c-0.4,0.1-0.7,0.3-1,0.6s-0.5,0.6-0.6,1c-0.2,0.4-0.2,0.9-0.2,1.4c0,0.5,0.1,0.9,0.2,1.3s0.3,0.8,0.6,1.1s0.6,0.5,1,0.7
+      c0.4,0.2,0.9,0.3,1.4,0.3c0.3,0,0.6,0,0.8,0s0.5-0.1,0.8-0.1s0.5-0.1,0.7-0.2c0.2-0.1,0.5-0.2,0.7-0.3l0.3,2.1
+      c-0.2,0.1-0.4,0.1-0.6,0.2s-0.5,0.1-0.8,0.2s-0.6,0.1-1,0.1C36.8,68,36.5,68,36.1,68z M37.4,67.5v-4.9h2.2v4.8L37.4,67.5z")
+    g#Fs
+      path.st5(d="M30.1,114c-19.6,5.2-27.5-24.3-7.9-29.6S49.7,108.8,30.1,114z")
+      g
+        path.st3(d="M17.9,104.5v-10h6.5v1.9h-5.2l0.9-1.6v9.6H17.9z M19.6,100.5v-1.8h4.6v1.8H19.6z")
+        path.st3(d="M25.2,102.7v-1.5h8.7v1.5H25.2z M25.9,98.6v-1.5h8.7v1.5H25.9z M26.5,104.5c0.1-0.5,1.6-9,1.7-9.3h1.6
+        c-0.1,0.3-1.6,8.8-1.7,9.3H26.5z M30,104.5c0.1-0.5,1.6-9,1.7-9.3h1.6c-0.1,0.3-1.6,8.8-1.7,9.3H30z")
+    g#F
+      path.st6(d="M46.8,146.6c-14.3,14.3-36-7.3-21.6-21.7S61.1,132.2,46.8,146.6z")
+      path.st3(d="M32.9,141.1v-10h6.5v1.9h-5.2l0.9-1.6v9.6H32.9z M34.6,137.2v-1.8h4.6v1.8H34.6z")
+    g#E
+      path.st7(d="M77.5,166.4c-5.2,19.6-34.8,11.7-29.6-7.9S82.7,146.9,77.5,166.4z")
+      path.st3(d="M59.5,168v-10h6.6v1.9h-5.5l1.1-1.5v9.3l-1-1.5h5.4v1.9H59.5z M61.2,163.8v-1.8h4.7v1.8H61.2z")
+    g#Ds
+      path.st8(d="M114,168.3c5.2,19.6-24.3,27.5-29.6,7.9S108.8,148.7,114,168.3z")
+      g
+        path.st3(d="M90.2,178.4v-10h4c0.8,0,1.4,0.1,2,0.3c0.6,0.2,1.1,0.5,1.5,1c0.4,0.4,0.7,0.9,0.9,1.6
+        c0.2,0.6,0.3,1.3,0.3,2.1c0,0.7-0.1,1.4-0.3,2s-0.5,1.1-0.9,1.6s-1,0.8-1.6,1.1c-0.7,0.3-1.5,0.4-2.5,0.4H90.2z M92.5,176.6h1.1
+        c0.7,0,1.3-0.1,1.7-0.4c0.4-0.3,0.7-0.7,0.9-1.1s0.3-1,0.3-1.6c0-0.5-0.1-0.9-0.2-1.3c-0.1-0.4-0.3-0.7-0.5-1
+        c-0.2-0.3-0.5-0.5-0.9-0.6c-0.4-0.1-0.8-0.2-1.3-0.2h-1.3V176.6z")
+        path.st3(d="M99.7,176.6V175h8.7v1.5H99.7z M100.4,172.5V171h8.7v1.5H100.4z M101,178.4c0.1-0.5,1.6-9,1.7-9.3h1.6
+        c-0.1,0.3-1.6,8.8-1.7,9.3H101z M104.4,178.4c0.1-0.5,1.6-9,1.7-9.3h1.6c-0.1,0.3-1.6,8.8-1.7,9.3H104.4z")
+    g#D
+      path.st9(d="M146.6,151.7c14.3,14.3-7.3,36-21.7,21.6S132.2,137.3,146.6,151.7z")
+      path.st3(d="M132.1,168v-10h4c0.8,0,1.4,0.1,2,0.3s1.1,0.5,1.5,1s0.7,0.9,0.9,1.6s0.3,1.3,0.3,2.1c0,0.7-0.1,1.4-0.3,2
+      c-0.2,0.6-0.5,1.1-0.9,1.6c-0.4,0.5-1,0.8-1.6,1.1c-0.7,0.3-1.5,0.4-2.5,0.4H132.1z M134.4,166.1h1.1c0.7,0,1.3-0.1,1.7-0.4
+      c0.4-0.3,0.7-0.7,0.9-1.1c0.2-0.5,0.3-1,0.3-1.6c0-0.5-0.1-0.9-0.2-1.3s-0.3-0.7-0.5-1s-0.5-0.5-0.9-0.6c-0.4-0.1-0.8-0.2-1.3-0.2
+      h-1.3V166.1z")
+    g#Cs
+      path.st10(d="M166.4,121c19.6,5.3,11.7,34.8-7.9,29.6S146.9,115.7,166.4,121z")
+      g
+        path.st3(d="M158.9,141.7c-0.8,0-1.5-0.1-2.2-0.4s-1.2-0.6-1.7-1.1c-0.5-0.5-0.8-1-1.1-1.6c-0.2-0.6-0.4-1.3-0.4-2.1
+        c0-1,0.2-1.9,0.6-2.7s1-1.4,1.8-1.9s1.7-0.7,2.7-0.7c0.3,0,0.6,0,1,0.1c0.3,0,0.6,0.1,0.9,0.2c0.3,0.1,0.6,0.2,0.8,0.2
+        c0.2,0.1,0.5,0.2,0.7,0.3l-0.2,2.1c-0.4-0.2-0.7-0.4-1-0.5s-0.7-0.2-1-0.3c-0.3-0.1-0.6-0.1-0.9-0.1c-0.6,0-1.1,0.1-1.5,0.4
+        c-0.4,0.2-0.8,0.6-1,1.1c-0.3,0.5-0.4,1.1-0.4,1.8c0,0.4,0.1,0.9,0.2,1.3c0.1,0.4,0.3,0.7,0.6,1s0.6,0.5,1,0.7
+        c0.4,0.2,0.9,0.3,1.4,0.3c0.2,0,0.4,0,0.7,0s0.4-0.1,0.7-0.1s0.4-0.2,0.7-0.3c0.2-0.1,0.4-0.3,0.7-0.4l0.1,2.1
+        c-0.2,0.1-0.4,0.2-0.7,0.3c-0.2,0.1-0.5,0.2-0.8,0.2c-0.3,0.1-0.5,0.1-0.8,0.1C159.4,141.7,159.1,141.7,158.9,141.7z")
+        path.st3(d="M162.8,139.7v-1.5h8.7v1.5H162.8z M163.5,135.7v-1.5h8.7v1.5H163.5z M164.1,141.5c0.1-0.5,1.6-9,1.7-9.3h1.6
+        c-0.1,0.3-1.6,8.8-1.7,9.3H164.1z M167.6,141.5c0.1-0.5,1.6-9,1.7-9.3h1.6c-0.1,0.3-1.6,8.8-1.7,9.3H167.6z")
+    g#C
+      path.st11(d="M168.3,84.4c19.6-5.2,27.5,24.3,7.9,29.6S148.7,89.7,168.3,84.4z")
+      path.st3(d="M172.6,104.5c-0.8,0-1.5-0.1-2.2-0.4s-1.2-0.6-1.7-1.1c-0.5-0.5-0.8-1-1.1-1.6c-0.2-0.6-0.4-1.3-0.4-2.1
+      c0-1,0.2-1.9,0.6-2.7c0.4-0.8,1-1.4,1.8-1.9c0.8-0.4,1.7-0.7,2.7-0.7c0.3,0,0.6,0,1,0.1c0.3,0,0.6,0.1,0.9,0.2
+      c0.3,0.1,0.6,0.2,0.8,0.2c0.2,0.1,0.5,0.2,0.7,0.3l-0.2,2.1c-0.4-0.2-0.7-0.4-1-0.5s-0.7-0.2-1-0.3c-0.3-0.1-0.6-0.1-0.9-0.1
+      c-0.6,0-1.1,0.1-1.5,0.4c-0.4,0.2-0.8,0.6-1,1.1c-0.3,0.5-0.4,1.1-0.4,1.8c0,0.4,0.1,0.9,0.2,1.3c0.1,0.4,0.3,0.7,0.6,1
+      s0.6,0.5,1,0.7c0.4,0.2,0.9,0.3,1.4,0.3c0.2,0,0.4,0,0.7,0s0.4-0.1,0.7-0.1s0.4-0.2,0.7-0.3c0.2-0.1,0.4-0.3,0.7-0.4l0.1,2.1
+      c-0.2,0.1-0.4,0.2-0.7,0.3c-0.2,0.1-0.5,0.2-0.8,0.2c-0.3,0.1-0.5,0.1-0.8,0.1C173.2,104.5,172.9,104.5,172.6,104.5z")
+    g#B
+      path.st12(d="M151.7,51.8c14.3-14.3,36,7.3,21.6,21.7S137.3,66.2,151.7,51.8z")
+      path.st3(d="M163.1,67.4h-3.5v-10h4c0.7,0,1.3,0.1,1.7,0.2c0.5,0.1,0.8,0.3,1.1,0.5c0.3,0.2,0.5,0.4,0.6,0.7
+      c0.1,0.3,0.2,0.6,0.2,0.9c0,0.4-0.1,0.8-0.2,1.1c-0.2,0.3-0.4,0.6-0.7,0.8c-0.3,0.2-0.6,0.3-0.9,0.4v0.2c0.4,0.1,0.8,0.2,1.1,0.4
+      c0.3,0.2,0.6,0.4,0.7,0.8c0.2,0.3,0.3,0.7,0.3,1.2c0,0.4-0.1,0.8-0.2,1.1c-0.1,0.3-0.4,0.6-0.7,0.9c-0.3,0.2-0.8,0.4-1.3,0.6
+      C164.7,67.3,164,67.4,163.1,67.4z M161.9,65.6h1.5c0.4,0,0.7,0,1-0.1s0.4-0.2,0.6-0.4c0.1-0.2,0.2-0.4,0.2-0.7
+      c0-0.4-0.1-0.7-0.4-0.9c-0.3-0.2-0.7-0.3-1.2-0.3h-2.1v-1.6h2c0.3,0,0.6-0.1,0.8-0.2c0.2-0.1,0.4-0.3,0.5-0.4
+      c0.1-0.2,0.2-0.4,0.2-0.7c0-0.4-0.1-0.7-0.3-0.9c-0.2-0.2-0.6-0.3-1-0.3h-1.6V65.6z")
+    g#As
+      path.st13(d="M121,32c5.3-19.6,34.8-11.7,29.6,7.9S115.7,51.6,121,32z")
+      g
+        path.st3(d="M126.1,40.2c0.2-0.5,3.1-9.5,3.3-10h3c0.2,0.5,3.1,9.4,3.2,10h-2.4c-0.1-0.5-2.3-8-2.3-8h-0.2l-2.3,8H126.1z
+          M128.5,38.2l0.2-1.7h4.5l0.2,1.7H128.5z")
+        path.st3(d="M136.1,38.4v-1.5h8.7v1.5H136.1z M136.8,34.4v-1.5h8.7v1.5H136.8z M137.4,40.2c0.1-0.5,1.6-9,1.7-9.3h1.6
+        c-0.1,0.3-1.6,8.8-1.7,9.3H137.4z M140.9,40.2c0.1-0.5,1.6-9,1.7-9.3h1.6c-0.1,0.3-1.6,8.8-1.7,9.3H140.9z")
+    g#A
+      path.st14(d="M84.4,30.1c-5.2-19.5,24.3-27.5,29.6-7.9S89.7,49.7,84.4,30.1z")
+      path.st3(d="M94.5,30.1c0.2-0.5,3.1-9.5,3.3-10h3c0.2,0.5,3.1,9.4,3.2,10h-2.4c-0.1-0.5-2.3-8-2.3-8h-0.2l-2.3,8H94.5z
+        M96.9,28l0.2-1.7h4.5l0.2,1.7H96.9z")
+</template>
+
+<style lang="postcss" scoped>
+.st0 {
+  fill: #333132;
+}
+
+.st1 {
+  fill: #F1F2F2;
+}
+
+.st2 {
+  fill: #E21A58;
+}
+
+.st3 {
+  fill: #FFFFFF;
+}
+
+.st4 {
+  fill: #EB0D8C;
+}
+
+.st5 {
+  fill: #922A8E;
+}
+
+.st6 {
+  fill: #2E3690;
+}
+
+.st7 {
+  fill: #0D73BB;
+}
+
+.st8 {
+  fill: #2BACE2;
+}
+
+.st9 {
+  fill: #00A99D;
+}
+
+.st10 {
+  fill: #06A64F;
+}
+
+.st11 {
+  fill: #88C541;
+}
+
+.st12 {
+  fill: #FCEE23;
+}
+
+.st13 {
+  fill: #F89521;
+}
+
+.st14 {
+  fill: #EF372F;
+}
+</style>
