+++ content/components.d.ts
@@ -7,8 +7,8 @@ export {}
 
 declare module 'vue' {
   export interface GlobalComponents {
-    AbcEditor: typeof import('./../components/abc/editor.vue')['default']
-    AbcRender: typeof import('./../components/abc/render.vue')['default']
+    AbcEditor: typeof import('./../components/abc/AbcEditor.vue')['default']
+    AbcRender: typeof import('./../components/abc/AbcRender.vue')['default']
     AbcState: typeof import('./../components/abc/state.ts')['default']
     AbcTunes: typeof import('./../components/abc/tunes.ts')['default']
     AuthorCard: typeof import('./../theme/components/author/card.vue')['default']
@@ -67,8 +67,8 @@ declare module 'vue' {
     ControlStart: typeof import('./../components/control/start.vue')['default']
     Counter: typeof import('./../theme/components/global/counter.vue')['default']
     Draggable: typeof import('./../theme/components/global/draggable.vue')['default']
+    Draw: typeof import('./../theme/components/draw/draw.js')['default']
     DrawControls: typeof import('./../theme/components/draw/controls.vue')['default']
-    DrawDraw: typeof import('./../theme/components/draw/draw.js')['default']
     DrawLayer: typeof import('./../theme/components/draw/layer.vue')['default']
     EditSeen: typeof import('./../theme/components/edit/seen.vue')['default']
     FullScreen: typeof import('./../theme/components/global/full-screen.vue')['default']
@@ -137,20 +137,20 @@ declare module 'vue' {
     ShopMessage: typeof import('./../theme/components/shop/message.vue')['default']
     ShopPrice: typeof import('./../theme/components/shop/price.vue')['default']
     SideBar: typeof import('./../theme/components/side/bar.vue')['default']
+    SideBarLevel: typeof import('./../theme/components/side/SideBarLevel.vue')['default']
     SidePanel: typeof import('./../theme/components/side/panel.vue')['default']
     SideSearch: typeof import('./../theme/components/side/search.vue')['default']
-    SideSideBarLevel: typeof import('./../theme/components/side/SideBarLevel.vue')['default']
     Spacer: typeof import('./../theme/components/global/spacer.vue')['default']
     StateDark: typeof import('./../theme/components/state/dark.vue')['default']
-    StateSound: typeof import('./../components/state/sound.vue')['default']
-    StateTransport: typeof import('./../components/state/transport.vue')['default']
-    SvgLogo: typeof import('./../components/svg/logo.vue')['default']
-    SvgRing: typeof import('./../components/svg/ring.vue')['default']
+    StateSound: typeof import('./../components/state/StateSound.vue')['default']
+    StateTransport: typeof import('./../components/state/StateTransport.vue')['default']
+    SvgLogo: typeof import('./../components/svg/SvgLogo.vue')['default']
+    SvgRing: typeof import('./../components/svg/SvgRing.vue')['default']
     SvgSave: typeof import('./../theme/components/global/svg-save.vue')['default']
-    SynthOscillators: typeof import('./../components/synth/oscillators.vue')['default']
-    SynthPanel: typeof import('./../components/synth/panel.vue')['default']
-    SynthSpeech: typeof import('./../components/synth/speech.vue')['default']
-    ToolsList: typeof import('./../components/tools/list.vue')['default']
+    SynthOscillators: typeof import('./../components/synth/SynthOscillators.vue')['default']
+    SynthPanel: typeof import('./../components/synth/SynthPanel.vue')['default']
+    SynthSpeech: typeof import('./../components/synth/SynthSpeech.vue')['default']
+    ToolsList: typeof import('./../components/tools/ToolsList.vue')['default']
     YoutubeEmbed: typeof import('./../theme/components/global/youtube-embed.vue')['default']
   }
 }
