+++ content/theory/resources/index.md
@@ -1,57 +1,8 @@
 ---
 title: Resources
 description: Links to some valuable information on music and color theory and more
-
+cover: ugur-akdemir.jpg
 date: 2019-02-02
-links:
-  - url: https://rhythmiclight.com/
-    title: Rhythmic light by Fred Collopy
-    description: The most full and honorable source of knowledge about light to sound correspondences. Started in 1998 and still growing!
-  - url: http://www.tonalsoft.com/enc/encyclopedia-index.aspx
-    title: Encyclopedia of Microtonal Music Theory
-    description: A very deep resource about all the subtleties and complexities of music theory by [Joe Monzo](http://www.tonalsoft.com/enc/m/monzo-writings.aspx)
-  - url: https://programmingdesignsystems.com/
-    title: A free digital book that teaches a practical introduction to the new foundations of graphic design.
-    description: By Rune Madsen
-  - url: https://mimo-international.com/MIMO/instrument-families.aspx
-    title: Musical instrument museums online
-    description: Explore the world collections of musical instruments
-  - url: http://www.musanim.com/
-    title: Music animation machine
-    description: A life long project of Stephen Malinowski – he manually animates music pieces into nice visual playalongs. He uses blue color for the tonic (C) and accending colors for every next step of a fifth.
-  - url: https://sethares.engr.wisc.edu/index.html
-    title: Home page of William A. Sethares
-    description: A very important figure in sensory dissonance research. His book "Tuning, Timbre, Spectrum, Scale" gives a comprehensive understanding of many important music concepts in a scientifically based way.
-  - url: https://www.acs.psu.edu/drussell/demos.html
-    title: Acoustics and Vibration Animations by Dr. Daniel A. Russell
-    description: Teaching Professor of Acoustics, Graduate Program in Acoustics, The Pennsylvania State University. He created animations illustrating acoustics and vibration, waves and oscillation concepts.
-  - url: http://www.phys.unsw.edu.au/music/
-    title: Music acoustics
-    description: Physics and music have been related for millenia. The art and science of music acoustics are presented here, in musician-friendly format, as is our research in music science.
-  - url: https://cielab.xyz/colorconv/
-    title: CIELAB.XYZ
-    description: Online color converter for different color models
-  - url: https://en.wikipedia.org/wiki/List_of_pitch_intervals
-    title: List of pitch intervals
-    description: Comprehesive Wikipedia collection of intervals
-  - url: https://en.wikipedia.org/wiki/List_of_chords
-    title: List of chords
-    description: Comprehesive Wikipedia collection of chords
-  - url: https://en.wikipedia.org/wiki/List_of_chord_progressions
-    title: List of chord progressions
-    description: Comprehesive Wikipedia collection of chord progressions
-  - url: https://en.wikipedia.org/wiki/List_of_musical_scales_and_modes
-    title: List of scales and modes
-    description: Comprehesive Wikipedia collection of pitch class collections
-  - url: http://musicnotation.org/systems/
-    title: The Music notation project
-    description: Many people struggle to learn to read and play music, and many give up before they become proficient. Could a better notation system make reading, writing, and playing music more enjoyable and easier to learn? We think so.
-  - url: https://ianring.com/musictheory/scales/
-    title: The Exciting Universe Of Music Theory
-    description: A complete study of all Scales in the existence by Ian Ring. "The place for all you music theory nerds to geek out. Bask in the warm bath of wisdom, and be envied by all your peers with your deep knowledge of musical lore. "
-  - url: http://allthescales.org/
-    title: AllTheScales.org
-    description: Based on the Western system of notes (as exemplified by the piano keyboard), there are 1490 possible scales. 2 In other words, with the combined genius of medieval monks through Beethoven through Stravinsky and Coltrane, we've managed to explore roughly 1% of this musical terrain.
 ---
